+++ Sources/Node/Convertible/Convertible+Node.swift
@@ -1,0 +1,9 @@
+extension Node: NodeConvertible { // Can conform to both if non-throwing implementations
+    public init(with node: Node, in context: Context) {
+        self = node
+    }
+
+    public func toNode() -> Node {
+        return self
+    }
+}
