+++ Sources/Node/Core/Node+Exports.swift
@@ -1,0 +1,3 @@
+// These are critical components to Node and are being exported
+@_exported import Polymorphic
+@_exported import PathIndexable
