--- Sources/Node/Utility/Logging.swift
@@ -1,23 +1,0 @@
-//
-//  Genome
-//
-//  Created by Logan Wright
-//  Copyright © 2016 lowriDevs. All rights reserved.
-//
-//  MIT
-//
-
-public var loggers: [(ErrorProtocol) -> Void] = [defaultLogger]
-
-private func defaultLogger(error: ErrorProtocol) {
-    print("Node.Error.\(error)")
-}
-
-extension ErrorProtocol {
-    func logged() -> Self {
-        loggers.forEach {
-            $0(self)
-        }
-        return self
-    }
-}
